libvirt: add flag to specify the bootloader to use
Using libvirt + xen, if libvirt is using libxl (the default toolstack
for Xen, since xend is deprecated from Xen 4.1) the bootloader is
harcoded to "/usr/bin/pygrub" [1]. In Debian and Ubuntu pygrub is installed
in /usr/lib/xen-<version>/bin/pygrub. OpenStack does not specify a
bootloader in the 'libvirt.xml' files generated for each instance therefore,
booting a machine with libivrt + xen using libxenlight (libxl) in OpenStack
is not possible.

This change introduces a new flag [libvirt]/xen_bootloader_path that can
be used to specify the correct path to the bootloader.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1104695

Change-Id: I9b5ab2f9d54c73b013c2eacce79e55f5f8b986e5