virt: add method for obtaining the list of cached images
Some virt drivers use a image cache to store the downloaded images. This
commit introduces a "get_cached_images_sha1" in the base class so that we
have a formal method to obtain the list of the SHA1 of the image IDs
that are cached in that node, without accessing directly each if the
image managers.

Change-Id: I56b1251518f842fc4b7e8e04e404b5f2a16a953f